Allow the admin node to act as a SNAT translator for BMCs.
This commit adds two things that only take effect if the BMC network
is not on the same subnet as the admin network:

* On the admin node, iptables rules are put into plce to allow
forwarding from the admin net to the BMC net and to snat any packets
so forwarded.  The admin node will create a vlan device that should
enable connectivity to the BMC network.

* On all other nodes, a route is added that will forward any packets
destined for the bmc network range to the admin node.

This should allow bmc and wsman calls over IP to work normally when
the client nodes are not on the same network as the bmcs.