Fix vulnerability caused by predictable pathnames for tcpdump packet captures.
Packet captures are cached from previous runs; however this requires
the use of predictable pathnames.  To prevent this becoming a security
risk, we create a dedicated directory and ensure that we own it and
it's not writable by anyone else.

See https://bugzilla.novell.com/show_bug.cgi?id=774967

Conflicts:

	chef/cookbooks/ohai/files/default/plugins/crowbar.rb