Don't try to install rubygem-cstruct
The ohai cookbook is also executed inside the sledgehammer image were it is not
possible to install packages (rpm/zypper is not present inside the opensuse
based sledgehammer). This needs to be solved in different way. E.g. by making
rubygem-cstruct part of sledgehammer on openSUSE and by adding  the gem to the
AutoYaST profile so that is installed during the base OS deployment.