Fix the way how pci bus ids are compared
Crowbar used to strip everything after the first '.' in the busid. So for a
node with two nics and busids of:
0000:00/0000:00:03.0/0000:01:02.0 and 0000:00/0000:00:03.0/0000:01:03.0
Crowbar would strip down the busids of both devices to: 0000:00/0000:00:03.
This means it was not possible to define a custom busorder for such a setup.

This commit changes the code to compare all components of the busids instead
stripping off some parts. It tries to be backwards compatible. If the busid in
the network proposal don't contain any '.' it falls back to the old method of
comparing the ids.

Unfortunately this code is duplicated in the deployer barclamp's cookbooks
and in the NodeObject class in the crowbar barclamp (added a note about that to
the code, as there is no easy way to get rid of the duplication).