Install right version of cstruct rubygem before calling ohai
This ensures that the cstruct rubygem is installed, even after we
upgraded to a new version of ruby.

We do not do this when running inside sledgehammer since the package is
already there, and due to size optimization of the image, the check a
package being installed is broken.

Note that this should actually help get rid of the hard-coding of this
gem in the provisioner barclamp.